[#2273] Prepare base classes for implementing Kafka based command router
