+++ clients/kafka-common/pom.xml
@@ -64,7 +64,11 @@
       </exclusions>
     </dependency>
     <dependency>
-      <groupId>io.vertx</groupId>
+      <groupId>org.eclipse.hono</groupId>
+            <artifactId>hono-client</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.vertx</groupId>
       <artifactId>vertx-kafka-client</artifactId>
     </dependency>
